Add geo_shape support to geo_centroid aggregation (#46299)
* Add geo_shape support to geo_centroid aggregation

This commit annotates geometries with centroid information
at index-time. Then each shapes centroid is used as input to
the geo_centroid aggregator so that it effectively calculates
the centroid of centroids of shapes within each bucket.

The centroid of each geo_shape field is calculated as the
centroid of all the points of all the shapes in the geometry
collection.

The centroid of each shape is calculated the same way geo_centroid
was -- using a kahan-summation to reduce precision loss.

* add centroid test