Add an index-time k-means implementation.
NOTE: this is an early prototype and is very hacky.

This commit adds a new codec KMeansCodec that performs k-means clustering during
indexing. When writing docvalues for the vector field, the codec will also
perform k-means and report certain metrics about the result. Currently the
result of the k-means is not persisted anywhere, and it performs k-means from
scratch every time a new segment is created.

Some notes:
* The original field must be called 'vector'.
* The mapping param 'iters' specifies the number of iterations of Lloyd's
  algorithm to run.
* The mapping param 'sample_fraction' specifies the fraction of the data to
  consider on each iteration.
* There is also a 'streaming' or sequential version of k-means, which updates
  the centroid locations every time a point is assigned to a cluster. It can be
  enabled by enabling the 'streaming' mapping param.