Report search progress back to main search task (#46309)
Each shard operation returns a TaskInfo object back to its main search task running in the coordinating node, together with the actual result of the operation. The status of the main search task can then include a summary of all the phases that it has run, which phase is currently being run, and for each shard operation there is a TaskInfo that allows to retrieve more info on the specific shard operation.